Basic enemy movement & Health loss & Growing
Basic enemy movement:
The enemy will move in a random direction for X amount of time then change.

Health Loss:
When the player touches an enemy their health goes down by 1 (starts a 3) and they cannot take damage for 2 seconds (shown by a blinking effec). The game restarts if their health is 0

Growing:
The players scale and the camers size change together, but the camera waits X amount of size increases to change