Eating
When the player touches an enemy that is larger the player will take damage. If the enemy is smaller the player will increase their size by 'enemySize/5' and the enemy will be destroyed.
It looks kind of strange atm with the enemies just dissapearing so maybe there can be an animation in a future version.
Also removed 'Volcanicling.psd'