Enemy spawning v2
Background music:
Added the audio file 'Background.wav' as a temp file until we find a better one.
The AudioController is a class that can be used in the future for controlling the audio and is attatched to an object that will exist in all scenes.

Enemy spawning:
Moved some thing in to functions so they can be used multiple times.
There is a limit on how many enemies can spawn and the limit increases every 5 seconds.
There will always be X% of enemies that are smaller than the player.
Added options for enemy speed to the enemy prefab in the editor.
When an enemy is eaten they will be moved to a different location and their speed will be randomized.

Others:
Moved the list of enemies from GameInfo to Spawning.
Fixed a bug that caused background chunks to constantly switch positions.
Increased the amount of background chunks to 35 so there is no pop-in when the player is a large size.