Radar
There is now a radar at the bottom of the screen. The radar shows enemies that are close to the player and how big they are. The players size is also shown, which makes it easier to know if you can eat an enemy or not.

There is a range option in the editor.
A new 'blip' is created every time a new enemy spawns.