Joystick style controls
The player can drag from any point on the screen to control the player. It works in the same way the old style does.

A large circle is placed at the ancor point and a smaller circle under the players finger.
There is an option to switch between the old and new controls in the editor.