Paired lazer enemies
This enemy is two basic enemies linked together that fire a beam between eachother every few seconds.
When the player hits the beam they take damage, but nothing happens when they hit one of the paired enemies.

The enemy scripts have been moved in to their own sub folder.
The linked enemies fly backwards because the sprite is in the other direction to the basic enemy.
When only 1 of the 2 linked enemies are on the radar and they fire the beam, the radar beam will fire down