Enemy revamp, joystick controls & intelligent enemy
All enemies will have the 'Enemy' script which controls the speed, size and pointing direction.
To control the enemy movement a new script "%type%_behaviour" will be created for each type of enemy. This script will control what the enemy does and how it moves.
Changed all other scripts to support the new system

The player can drag from any point on the screen to control the player. It works in the same way the old style does.

A large circle is placed at the ancor point and a smaller circle under the players finger.
There is an option to switch between the old and new controls in the editor. In the code you need to get the 'Player' component from the player and do "Player.enableJoystick = <true / false>"

The intelligent enemy will act the same as the basic enemy until it is close enugh to the player. When in range it will move towards the player if it is larger and move away if it is smaller.
When the enemy respawns it's speed will be changed to an amount based on how long the game has been played for. It currently takes around 1min 40sec to reach max speed. It is slow to start so maybe set the min speed higher, but this will decrease the time it takes to get to the max speed.

Added 'speedMultiplier' to the player for use with future enemies and powerups.
Spawing now takes an array of GameObject's for the enemies and selects a random one when spawning.
Created a sub folder for enemy scripts.
Changed the enemy prefab in Michael's folder so the new system works in the Main scene.